ragwith/requirements.txt
langchain>=0.1.12
langchain-community>=0.0.27
langchain-ollama>=0.3.1
streamlit>=1.32.0
duckdb>=0.10.0
sentence-transformers>=2.5.1
python-dotenv>=1.0.1
pandas>=2.2.1
numpy>=1.26.4